requirements.txt
# Xinch Dependencies
# Cross-chain USDC swaps between Ethereum and XRPL

# Ethereum dependencies
web3>=6.0.0
eth-account>=0.9.0

# XRPL dependencies  
xrpl-py>=2.0.0

# Configuration and utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Data handling
pydantic>=2.0.0

# Optional: For advanced crypto operations
cryptography>=3.4.8 